+++ lib/src/widgets/bottom_navigation_builder.dart
@@ -26,9 +26,26 @@ class BottomNavigationBuilder implements NavigatorBuilder {
   /// Creates a [BottomNavigationBuilder] instance.
   ///
   const BottomNavigationBuilder({
-    required this.bottomNavigationItems,
+    this.bottomNavigationItems = const <BottomNavigationBarItem>[],
     this.parameters = const BottomNavigationBarParameters(),
-  });
+    this.navigationBarItems = const <NavigationDestination>[],
+    this.navigationBarParameters = const NavigationBarParameters(),
+    bool? material3,
+  }) : _material3 = material3 ?? false;
+
+  /// Creates a [BottomNavigationBuilder] instance that uses Material 3 [NavigationBar]
+  /// widget.
+  ///
+  factory BottomNavigationBuilder.navigationBar({
+    required List<NavigationDestination> navigationBarItems,
+    NavigationBarParameters? navigationBarParameters,
+  }) =>
+      BottomNavigationBuilder(
+        navigationBarItems: navigationBarItems,
+        navigationBarParameters:
+            navigationBarParameters ?? const NavigationBarParameters(),
+        material3: true,
+      );
 
   /// A list of [BottomNavigationBarItems], that corresponds to the navigator's
   /// destination list.
@@ -46,16 +63,37 @@ class BottomNavigationBuilder implements NavigatorBuilder {
   ///
   final BottomNavigationBarParameters parameters;
 
+  /// A list of [NavigationDestination] widgets, that corresponds to the navigator's
+  /// destination list.
+  ///
+  /// The list must contain the same number of items, following with the same order
+  /// as a destination list specified in the navigator.
+  ///
+  final List<NavigationDestination> navigationBarItems;
+
+  /// A set of [NavigationBar] parameters.
+  ///
+  /// Contains all supported parameters to customize [NavigationBar] widget.
+  /// Doesn't include 'items', 'onTap' and 'currentIndex', which are managed by
+  /// [BottomNavigationBuilder].
+  ///
+  final NavigationBarParameters navigationBarParameters;
+
+  final bool _material3;
+
   @override
   Widget build(BuildContext context, NavigationController navigator) {
     final currentDestination = navigator.currentDestination;
     return _BottomNavigationWrapper(
       destination: currentDestination,
-      items: bottomNavigationItems,
-      parameters: parameters,
       onSelectBottomTab: (index) =>
           navigator.goTo(navigator.destinations[index]),
       selectedIndex: navigator.destinations.indexOf(currentDestination),
+      items: bottomNavigationItems,
+      parameters: parameters,
+      navigationBarItems: navigationBarItems,
+      navigationBarParameters: navigationBarParameters,
+      material3: _material3,
     );
   }
 }
@@ -64,11 +102,27 @@ class _BottomNavigationWrapper extends StatefulWidget {
   const _BottomNavigationWrapper({
     Key? key,
     required this.destination,
-    required this.items,
     required this.onSelectBottomTab,
     required this.selectedIndex,
-    required this.parameters,
+    this.items = const <BottomNavigationBarItem>[],
+    this.parameters = const BottomNavigationBarParameters(),
+    this.navigationBarItems = const <NavigationDestination>[],
+    this.navigationBarParameters = const NavigationBarParameters(),
+    this.material3 = false,
   }) : super(key: key);
+
+  factory _BottomNavigationWrapper.navigationBar({
+    required Destination destination,
+    required List<NavigationDestination> navigationBarItems,
+    required int selectedIndex,
+    required void Function(int) onSelectBottomTab,
+  }) =>
+      _BottomNavigationWrapper(
+        destination: destination,
+        navigationBarItems: navigationBarItems,
+        onSelectBottomTab: onSelectBottomTab,
+        selectedIndex: selectedIndex,
+      );
 
   final Destination destination;
 
@@ -79,6 +133,12 @@ class _BottomNavigationWrapper extends StatefulWidget {
   final int selectedIndex;
 
   final BottomNavigationBarParameters parameters;
+
+  final List<NavigationDestination> navigationBarItems;
+
+  final NavigationBarParameters navigationBarParameters;
+
+  final bool material3;
 
   @override
   _BottomNavigationWrapperState createState() =>
@@ -101,35 +161,49 @@ class _BottomNavigationWrapperState extends State<_BottomNavigationWrapper> {
       builder: (context) => Scaffold(
         body: Stack(
           children: [
-            ..._content.entries.map((entry) => Offstage(
-              offstage: _indexes[entry.key] != widget.selectedIndex,
-              child: entry.value,
-            )).toList(),
+            ..._content.entries
+                .map((entry) => Offstage(
+                      offstage: _indexes[entry.key] != widget.selectedIndex,
+                      child: entry.value,
+                    ))
+                .toList(),
           ],
         ),
-        bottomNavigationBar: BottomNavigationBar(
-          items: widget.items,
-          currentIndex: widget.selectedIndex,
-          onTap: widget.onSelectBottomTab,
-          elevation: widget.parameters.elevation,
-          type: widget.parameters.type,
-          fixedColor: widget.parameters.fixedColor,
-          backgroundColor: widget.parameters.backgroundColor,
-          iconSize: widget.parameters.iconSize,
-          selectedItemColor: widget.parameters.selectedItemColor,
-          unselectedItemColor: widget.parameters.unselectedItemColor,
-          selectedIconTheme: widget.parameters.selectedIconTheme,
-          unselectedIconTheme: widget.parameters.unselectedIconTheme,
-          selectedFontSize: widget.parameters.selectedFontSize,
-          unselectedFontSize: widget.parameters.unselectedFontSize,
-          selectedLabelStyle: widget.parameters.selectedLabelStyle,
-          unselectedLabelStyle: widget.parameters.unselectedLabelStyle,
-          showSelectedLabels: widget.parameters.showSelectedLabels,
-          showUnselectedLabels: widget.parameters.showUnselectedLabels,
-          mouseCursor: widget.parameters.mouseCursor,
-          enableFeedback: widget.parameters.enableFeedback,
-          landscapeLayout: widget.parameters.landscapeLayout,
-        ),
+        bottomNavigationBar: widget.material3
+            ? NavigationBar(
+                animationDuration:
+                    widget.navigationBarParameters.animationDuration,
+                selectedIndex: widget.selectedIndex,
+                destinations: widget.navigationBarItems,
+                onDestinationSelected: widget.onSelectBottomTab,
+                backgroundColor: widget.navigationBarParameters.backgroundColor,
+                elevation: widget.navigationBarParameters.elevation,
+                height: widget.navigationBarParameters.height,
+                labelBehavior: widget.navigationBarParameters.labelBehavior,
+              )
+            : BottomNavigationBar(
+                items: widget.items,
+                currentIndex: widget.selectedIndex,
+                onTap: widget.onSelectBottomTab,
+                elevation: widget.parameters.elevation,
+                type: widget.parameters.type,
+                fixedColor: widget.parameters.fixedColor,
+                backgroundColor: widget.parameters.backgroundColor,
+                iconSize: widget.parameters.iconSize,
+                selectedItemColor: widget.parameters.selectedItemColor,
+                unselectedItemColor: widget.parameters.unselectedItemColor,
+                selectedIconTheme: widget.parameters.selectedIconTheme,
+                unselectedIconTheme: widget.parameters.unselectedIconTheme,
+                selectedFontSize: widget.parameters.selectedFontSize,
+                unselectedFontSize: widget.parameters.unselectedFontSize,
+                selectedLabelStyle: widget.parameters.selectedLabelStyle,
+                unselectedLabelStyle: widget.parameters.unselectedLabelStyle,
+                showSelectedLabels: widget.parameters.showSelectedLabels,
+                showUnselectedLabels: widget.parameters.showUnselectedLabels,
+                mouseCursor: widget.parameters.mouseCursor,
+                enableFeedback: widget.parameters.enableFeedback,
+                landscapeLayout: widget.parameters.landscapeLayout,
+              ),
       ),
     );
   }
@@ -138,13 +212,15 @@ class _BottomNavigationWrapperState extends State<_BottomNavigationWrapper> {
   void didUpdateWidget(_BottomNavigationWrapper oldWidget) {
     super.didUpdateWidget(oldWidget);
     bool needsRebuild = false;
+    if (widget.material3 != oldWidget.material3) {
+      needsRebuild = true;
+    }
     if (!widget.destination.isFinalDestination) {
       needsRebuild = true;
       _content[widget.destination] = widget.destination.build(context);
       _indexes[widget.destination] = widget.selectedIndex;
-    }
-    else if (oldWidget.selectedIndex != widget.selectedIndex
-              && !_content.containsKey(widget.destination)) {
+    } else if (oldWidget.selectedIndex != widget.selectedIndex &&
+        !_content.containsKey(widget.destination)) {
       needsRebuild = true;
       _content[widget.destination] = widget.destination.build(context);
       _indexes[widget.destination] = widget.selectedIndex;
@@ -268,4 +344,45 @@ class BottomNavigationBarParameters {
   /// [BottomNavigationBar.landscapeLayout]
   ///
   final BottomNavigationBarLandscapeLayout? landscapeLayout;
+}
+
+/// Contains parameters to customize the [NavigationBar].
+///
+/// It includes all the same arguments as the [NavigationBar()], excepting
+/// the 'items', 'onTap' and 'currentIndex', which are managed by the [BottomNavigationBuilder].
+///
+/// See also:
+/// - [BottomNavigationBuilder]
+/// - [NavigationBar]
+///
+class NavigationBarParameters {
+  /// Create a [NavigationBarParameters] instance.
+  ///
+  const NavigationBarParameters({
+    this.animationDuration,
+    this.backgroundColor,
+    this.elevation,
+    this.height,
+    this.labelBehavior,
+  });
+
+  /// [NavigationBar.animationDuration]
+  ///
+  final Duration? animationDuration;
+
+  /// [NavigationBar.backgroundColor]
+  ///
+  final Color? backgroundColor;
+
+  /// [NavigationBar.elevation]
+  ///
+  final double? elevation;
+
+  /// [NavigationBar.height]
+  ///
+  final double? height;
+
+  /// [NavigationBar.labelBehavior]
+  ///
+  final NavigationDestinationLabelBehavior? labelBehavior;
 }